fix(test/xml): moved functions with pandas depency from __write_aux to __write_structure_specific_aux
